+++ desktop/php/060_AbeilleReHomePart.php
@@ -29,7 +29,7 @@
 		<br/>
 		<select name="zigateY" style="width : 40%">
 		<?php
-			for ( $i=1; $i<=config::byKey('zigateNb', 'Abeille', '0', 1); $i++ ) {
+			for ( $i=1; $i<= maxNbOfZigate; $i++ ) {
 				echo '<option value="' . $i . '">Zigate' . $i . '</option>';
 			}
 		?>